+++ config/buymeacoffee.php
@@ -1,0 +1,6 @@
+<?php
+
+return [
+    'url' => env('BUY_ME_A_COFFEE_WEBHOOK_HANDLER_URL', '/webhooks/buymeacoffee'),
+    'secret' => env('BUY_ME_A_COFFEE_WEBHOOK_SECRET'),
+];