+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # Changelog
 
-All notable changes to `:package_name` will be documented in this file.
+All notable changes to `laravel-buymeacoffee-webhook-handler` will be documented in this file.
+
+## 0.1
+
+Initial release having middleware and controller in place. The events will be fired on incoming requests.